Add better handling of API errors

+++ app/controllers/application_controller.rb
@@ -8,6 +8,11 @@ class ApplicationController < ActionController::Base
   rescue_from JsonApiClient::Errors::NotAuthorized, with: :api_not_authorized_error
   rescue_from JsonApiClient::Errors::ConnectionError, with: :api_connection_error
   rescue_from JsonApiClient::Errors::ServerError, with: :api_server_error
+  rescue_from JsonApiClient::Errors::AccessDenied, with: :api_access_denied
+
+  rescue_from ActionView::Template::Error do |e|
+    rescue_with_handler(e.cause) or raise(e)
+  end
 
   private
 
@@ -40,9 +45,18 @@ class ApplicationController < ActionController::Base
     drop_session
   end
 
+  def api_access_denied
+    flash[:danger] = "Sorry, you don't have access to that feature."
+    drop_session
+  end
+
   def drop_session
-    session.delete(:api_key)
-    redirect_to :login
+    if request.path == root_path
+      session.delete(:api_key)
+      redirect_to :login
+    else
+      redirect_to :root
+    end
   end
 
 end
